+++ client/fingerprint/cgroup_test.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/nomad/client/config"
+	cstructs "github.com/hashicorp/nomad/client/structs"
 	"github.com/hashicorp/nomad/nomad/structs"
 )
 
@@ -39,64 +40,91 @@ func (m *MountPointDetectorEmptyMountPoint) MountPoint() (string, error) {
 }
 
 func TestCGroupFingerprint(t *testing.T) {
-	f := &CGroupFingerprint{
-		logger:             testLogger(),
-		lastState:          cgroupUnavailable,
-		mountPointDetector: &MountPointDetectorMountPointFail{},
+	{
+		f := &CGroupFingerprint{
+			logger:             testLogger(),
+			lastState:          cgroupUnavailable,
+			mountPointDetector: &MountPointDetectorMountPointFail{},
+		}
+
+		node := &structs.Node{
+			Attributes: make(map[string]string),
+		}
+
+		request := &cstructs.FingerprintRequest{Config: &config.Config{}, Node: node}
+		var response cstructs.FingerprintResponse
+		err := f.Fingerprint(request, &response)
+		if err == nil {
+			t.Fatalf("expected an error")
+		}
+
+		if a, _ := response.Attributes["unique.cgroup.mountpoint"]; a != "" {
+			t.Fatalf("unexpected attribute found, %s", a)
+		}
 	}
 
-	node := &structs.Node{
-		Attributes: make(map[string]string),
+	{
+		f := &CGroupFingerprint{
+			logger:             testLogger(),
+			lastState:          cgroupUnavailable,
+			mountPointDetector: &MountPointDetectorValidMountPoint{},
+		}
+
+		node := &structs.Node{
+			Attributes: make(map[string]string),
+		}
+
+		request := &cstructs.FingerprintRequest{Config: &config.Config{}, Node: node}
+		var response cstructs.FingerprintResponse
+		err := f.Fingerprint(request, &response)
+		if err != nil {
+			t.Fatalf("unexpected error, %s", err)
+		}
+		if a, ok := response.Attributes["unique.cgroup.mountpoint"]; !ok {
+			t.Fatalf("unable to find attribute: %s", a)
+		}
 	}
 
-	ok, err := f.Fingerprint(&config.Config{}, node)
-	if err == nil {
-		t.Fatalf("expected an error")
-	}
-	if ok {
-		t.Fatalf("should not apply")
-	}
-	if a, ok := node.Attributes["unique.cgroup.mountpoint"]; ok {
-		t.Fatalf("unexpected attribute found, %s", a)
-	}
+	{
+		f := &CGroupFingerprint{
+			logger:             testLogger(),
+			lastState:          cgroupUnavailable,
+			mountPointDetector: &MountPointDetectorEmptyMountPoint{},
+		}
 
-	f = &CGroupFingerprint{
-		logger:             testLogger(),
-		lastState:          cgroupUnavailable,
-		mountPointDetector: &MountPointDetectorValidMountPoint{},
-	}
+		node := &structs.Node{
+			Attributes: make(map[string]string),
+		}
 
-	node = &structs.Node{
-		Attributes: make(map[string]string),
+		request := &cstructs.FingerprintRequest{Config: &config.Config{}, Node: node}
+		var response cstructs.FingerprintResponse
+		err := f.Fingerprint(request, &response)
+		if err != nil {
+			t.Fatalf("unexpected error, %s", err)
+		}
+		if a, _ := response.Attributes["unique.cgroup.mountpoint"]; a != "" {
+			t.Fatalf("unexpected attribute found, %s", a)
+		}
 	}
+	{
+		f := &CGroupFingerprint{
+			logger:             testLogger(),
+			lastState:          cgroupAvailable,
+			mountPointDetector: &MountPointDetectorValidMountPoint{},
+		}
 
-	ok, err = f.Fingerprint(&config.Config{}, node)
-	if err != nil {
-		t.Fatalf("unexpected error, %s", err)
-	}
-	if !ok {
-		t.Fatalf("should apply")
-	}
-	assertNodeAttributeContains(t, node, "unique.cgroup.mountpoint")
+		node := &structs.Node{
+			Attributes: make(map[string]string),
+		}
 
-	f = &CGroupFingerprint{
-		logger:             testLogger(),
-		lastState:          cgroupUnavailable,
-		mountPointDetector: &MountPointDetectorEmptyMountPoint{},
-	}
-
-	node = &structs.Node{
-		Attributes: make(map[string]string),
-	}
-
-	ok, err = f.Fingerprint(&config.Config{}, node)
-	if err != nil {
-		t.Fatalf("unexpected error, %s", err)
-	}
-	if !ok {
-		t.Fatalf("should apply")
-	}
-	if a, ok := node.Attributes["unique.cgroup.mountpoint"]; ok {
-		t.Fatalf("unexpected attribute found, %s", a)
+		request := &cstructs.FingerprintRequest{Config: &config.Config{}, Node: node}
+		var response cstructs.FingerprintResponse
+		err := f.Fingerprint(request, &response)
+		if err != nil {
+			t.Fatalf("unexpected error, %s", err)
+		}
+		if a, _ := response.Attributes["unique.cgroup.mountpoint"]; a == "" {
+			t.Fatalf("expected attribute to be found, %s", a)
+		}
 	}
 }