+++ client/structs/structs.go
@@ -4,6 +4,9 @@ import (
 	"crypto/md5"
 	"io"
 	"strconv"
+
+	"github.com/hashicorp/nomad/client/config"
+	"github.com/hashicorp/nomad/nomad/structs"
 )
 
 // MemoryStats holds memory usage related stats
@@ -183,4 +186,66 @@ func (d *DriverNetwork) Hash() []byte {
 		io.WriteString(h, strconv.Itoa(v))
 	}
 	return h.Sum(nil)
+}
+
+// FingerprintRequest is a request which a fingerprinter accepts to fingerprint
+// the node
+type FingerprintRequest struct {
+	Config *config.Config
+	Node   *structs.Node
+}
+
+// FingerprintResponse is the response which a fingerprinter annotates with the
+// results of the fingerprint method
+type FingerprintResponse struct {
+	Attributes map[string]string
+	Links      map[string]string
+	Resources  *structs.Resources
+
+	// Detected is a boolean indicating whether the fingerprinter detected
+	// if the resource was available
+	Detected bool
+}
+
+// AddAttribute adds the name and value for a node attribute to the fingerprint
+// response
+func (f *FingerprintResponse) AddAttribute(name, value string) {
+	// initialize Attributes if it has not been already
+	if f.Attributes == nil {
+		f.Attributes = make(map[string]string, 0)
+	}
+
+	f.Attributes[name] = value
+}
+
+// RemoveAttribute sets the given attribute to empty, which will later remove
+// it entirely from the node
+func (f *FingerprintResponse) RemoveAttribute(name string) {
+	// initialize Attributes if it has not been already
+	if f.Attributes == nil {
+		f.Attributes = make(map[string]string, 0)
+	}
+
+	f.Attributes[name] = ""
+}
+
+// AddLink adds a link entry to the fingerprint response
+func (f *FingerprintResponse) AddLink(name, value string) {
+	// initialize Links if it has not been already
+	if f.Links == nil {
+		f.Links = make(map[string]string, 0)
+	}
+
+	f.Links[name] = value
+}
+
+// RemoveLink removes a link entry from the fingerprint response. This will
+// later remove it entirely from the node
+func (f *FingerprintResponse) RemoveLink(name string) {
+	// initialize Links if it has not been already
+	if f.Links == nil {
+		f.Links = make(map[string]string, 0)
+	}
+
+	f.Links[name] = ""
 }